move js code to deno_net
