+++ runtime/Cargo.toml
@@ -23,6 +23,7 @@ deno_console = { version = "0.10.0", path = "../extensions/console" }
 deno_core = { version = "0.91.0", path = "../core" }
 deno_crypto = { version = "0.24.0", path = "../extensions/crypto" }
 deno_fetch = { version = "0.32.0", path = "../extensions/fetch" }
+deno_net = { version = "0.1.0", path = "../extensions/net" }
 deno_timers = { version = "0.8.0", path = "../extensions/timers" }
 deno_url = { version = "0.10.0", path = "../extensions/url" }
 deno_web = { version = "0.41.0", path = "../extensions/web" }
