+++ runtime/js/01_errors.js
@@ -2,6 +2,9 @@
 "use strict";
 
 ((window) => {
+  const core = window.Deno.core;
+  const { BadResource, Interrupted } = core;
+
   class NotFound extends Error {
     constructor(msg) {
       super(msg);
@@ -86,13 +89,6 @@
     }
   }
 
-  class Interrupted extends Error {
-    constructor(msg) {
-      super(msg);
-      this.name = "Interrupted";
-    }
-  }
-
   class WriteZero extends Error {
     constructor(msg) {
       super(msg);
@@ -104,13 +100,6 @@
     constructor(msg) {
       super(msg);
       this.name = "UnexpectedEof";
-    }
-  }
-
-  class BadResource extends Error {
-    constructor(msg) {
-      super(msg);
-      this.name = "BadResource";
     }
   }
